+++ Source/Gapotchenko.FX/Properties/Resources.Designer.cs
@@ -19,7 +19,7 @@ namespace Gapotchenko.FX.Properties {
     // class via a tool like ResGen or Visual Studio.
     // To add or remove a member, edit your .ResX file then rerun ResGen
     // with the /str option, or rebuild your VS project.
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "15.0.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "16.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     internal class Resources {
@@ -102,6 +102,24 @@ namespace Gapotchenko.FX.Properties {
         internal static string ProgramExit_Message {
             get {
                 return ResourceManager.GetString("ProgramExit_Message", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Non-exhaustive switch expression failed to match its input..
+        /// </summary>
+        internal static string SwitchExpressionException_Message {
+            get {
+                return ResourceManager.GetString("SwitchExpressionException_Message", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Unmatched value was {0}..
+        /// </summary>
+        internal static string SwitchExpressionException_UnmatchedValue {
+            get {
+                return ResourceManager.GetString("SwitchExpressionException_UnmatchedValue", resourceCulture);
             }
         }